+++ packages/cli/package.json
@@ -40,6 +40,7 @@
   "dependencies": {
     "@platformatic/authenticate": "workspace:*",
     "@platformatic/db": "workspace:*",
+    "@platformatic/service": "workspace:*",
     "colorette": "^2.0.19",
     "commist": "^3.1.2",
     "desm": "^1.2.0",
