+++ packages/db/test/cli/validations.test.mjs
@@ -1,4 +1,4 @@
-import { cliPath } from './helper.mjs'
+import { cliPath } from './helper.js'
 import { test } from 'tap'
 import { join } from 'desm'
 import { readFile } from 'fs/promises'
